# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI-Powered Developer Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /repositories/{repoId}:
    get:
      operationId: getRepository
      tags: [repositories]
      summary: Get repository details
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
  /repositories/{repoId}/stats:
    get:
      operationId: getRepositoryStats
      tags: [repositories]
      summary: Get repository statistics summary
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repository stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryStats"
  /repositories/{repoId}/pull-requests:
    get:
      operationId: listPullRequests
      tags: [pullRequests]
      summary: List pull requests for a repository
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
            enum: [open, closed, all]
            default: open
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of pull requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PullRequest"
  /repositories/{repoId}/pull-requests/{prId}:
    get:
      operationId: getPullRequest
      tags: [pullRequests]
      summary: Get pull request details
      parameters:
        - name: repoId
          in: path
          required: true
          schema:
            type: integer
        - name: prId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pull request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequest"
...
components:
  schemas:
...
    Repository:
      type: object
      properties:
        id:
          type: integer
        githubId:
          type: string
        name:
          type: string
        fullName:
          type: string
        description:
          type: string
        language:
          type: string
        stars:
          type: integer
        forks:
          type: integer
        openPRs:
          type: integer
        lastAnalyzedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        reviewCount:
          type: integer
        latestRiskScore:
          type: number
          nullable: true
        latestHealthScore:
          type: number
          nullable: true
      required: [id, githubId, name, fullName, stars, forks, openPRs]
    RepositoryStats:
      type: object
      properties:
        repoId:
          type: integer
        totalCommits:
          type: integer
        totalPRs:
          type: integer
        mergedPRs:
          type: integer
        avgMergeTimeHours:
          type: number
        avgRiskScore:
          type: number
        bugFrequency:
          type: number
        topContributors:
          type: array
          items:
            type: object
            properties:
              login:
                type: string
              commits:
                type: integer
            required: [login, commits]
      required: [repoId, totalCommits, totalPRs, mergedPRs, avgMergeTimeHours, avgRiskScore, bugFrequency, topContributors]
    PullRequest:
      type: object
      properties:
        id:
          type: integer
        repoId:
          type: integer
        prNumber:
          type: integer
        title:
          type: string
        body:
          type: string
        state:
          type: string
          enum: [open, closed, merged]
        authorLogin:
          type: string
        authorAvatarUrl:
          type: string
        filesChanged:
          type: integer
        additions:
          type: integer
        deletions:
          type: integer
        riskScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high]
        hasReview:
          type: boolean
        reviewStatus:
          type: string
          enum: [pending, in_progress, completed]
        mergedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, repoId, prNumber, title, state, authorLogin, authorAvatarUrl, filesChanged, additions, deletions, hasReview]
...